+++ bash/create_db.sh
@@ -1,3 +1,4 @@
 rm -vf $ydb_dir/$ydb_rel/g/*.dat
+rm -vf $ydb_dir/$ydb_rel/g/*.mjl*
 mupip create
-mupip set -journal=disable -region '*'
+mupip set -journal=enable -region '*'