+++ src/tests/delete.nim
@@ -83,8 +83,7 @@ proc testKill3() =
   let s = getvar  ^X(1)
   assert "hello" == s
   killnode: ^X(1) # delete node
-  doAssertRaises(YdbError): # expect exception because node removed
-    discard getvar  ^X(1)
+  assert "" == getvar ^X(1)
   
   # create a tree
   setvar: ^X(1,1)="hello"
@@ -92,8 +91,7 @@ proc testKill3() =
   let dta = data: ^X(1) 
   assert 10 == dta # Expect no data but subtree
   kill: ^X(1)
-  doAssertRaises(YdbError): # expect exception because node removed
-    discard  getvar  ^X(1)
+  assert "" == getvar ^X(1)
 
 if isMainModule:
   test "killNode": testKill1()